feat: move granite formatter hf generation function to use llguidance instead of xgrammar
